requirements.txt
absl-py==2.1.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
comet_ml==3.49.5
configobj==5.0.9
contourpy==1.3.1
cycler==0.12.1
docopt==0.6.2
dulwich==0.22.8
everett==3.1.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
grpcio==1.71.0
higher==0.2.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pipreqs==0.4.13
protobuf==4.23.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
pyparsing==3.2.1
python-box==6.1.0
python-dateutil==2.9.0.post0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
scipy==1.15.2
semantic-version==2.10.0
sentry-sdk==2.24.0
setuptools==76.1.0
simplejson==3.20.1
six==1.17.0
standard-imghdr==3.13.0
sympy==1.13.1
tensorboard==2.15.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
wurlitzer==3.1.1
yarg==0.1.10
